move eids config to submodules
