+++ modules/growthCodeIdSystem.js
@@ -5,7 +5,7 @@
  * @requires module:modules/userId
  */
 
-import {logError, logInfo, tryAppendQueryString} from '../src/utils.js';
+import {logError, logInfo, pick, tryAppendQueryString} from '../src/utils.js';
 import {ajax} from '../src/ajax.js';
 import { submodule } from '../src/hook.js'
 import {getStorageManager} from '../src/storageManager.js';
@@ -166,6 +166,25 @@ export const growthCodeIdSubmodule = {
       }
     };
     return { callback: resp };
+  },
+  eids: {
+    'growthCodeId': {
+      getValue: function(data) {
+        return data.gc_id
+      },
+      source: 'growthcode.io',
+      atype: 1,
+      getUidExt: function(data) {
+        const extendedData = pick(data, [
+          'h1',
+          'h2',
+          'h3',
+        ]);
+        if (Object.keys(extendedData).length) {
+          return extendedData;
+        }
+      }
+    },
   }
 };
