+++ modules/oneKeyIdSystem.js
@@ -72,6 +72,32 @@ export const oneKeyIdSubmodule = {
     return {
       callback: getIdsAndPreferences
     };
+  },
+  eids: {
+    'oneKeyData': {
+      getValue: function(data) {
+        if (data && Array.isArray(data.identifiers) && data.identifiers[0]) {
+          return data.identifiers[0].value;
+        }
+      },
+      source: 'paf',
+      atype: 1,
+      getEidExt: function(data) {
+        if (data && data.preferences) {
+          return {preferences: data.preferences};
+        }
+      },
+      getUidExt: function(data) {
+        if (data && Array.isArray(data.identifiers) && data.identifiers[0]) {
+          const id = data.identifiers[0];
+          return {
+            version: id.version,
+            type: id.type,
+            source: id.source
+          };
+        }
+      }
+    }
   }
 };
