+++ modules/tapadIdSystem.js
@@ -56,6 +56,12 @@ export const tapadIdSubmodule = {
         );
       }
     }
+  },
+  eids: {
+    'tapadId': {
+      source: 'tapad.com',
+      atype: 1
+    },
   }
 }
 submodule('userId', tapadIdSubmodule);
