+++ modules/tncIdSystem.js
@@ -57,6 +57,12 @@ export const tncidSubModule = {
     return {
       callback: function (cb) { return tncCallback(cb); }
     }
+  },
+  eids: {
+    'tncid': {
+      source: 'thenewco.it',
+      atype: 3
+    },
   }
 }
