+++ modules/userId/eids.js
@@ -1,446 +1,7 @@
 import { pick, isFn, isStr, isPlainObject, deepAccess } from '../../src/utils.js';
 
 // Each user-id sub-module is expected to mention respective config here
-export const USER_IDS_CONFIG = {
-
-  // key-name : {config}
-
-  // GrowthCode
-  'growthCodeId': {
-    getValue: function(data) {
-      return data.gc_id
-    },
-    source: 'growthcode.io',
-    atype: 1,
-    getUidExt: function(data) {
-      const extendedData = pick(data, [
-        'h1',
-        'h2',
-        'h3',
-      ]);
-      if (Object.keys(extendedData).length) {
-        return extendedData;
-      }
-    }
-  },
-
-  // utiq
-  'utiq': {
-    source: 'utiq.com',
-    atype: 1,
-    getValue: function (data) {
-      return data;
-    },
-  },
-
-  // intentIqId
-  'intentIqId': {
-    source: 'intentiq.com',
-    atype: 1
-  },
-
-  // naveggId
-  'naveggId': {
-    source: 'navegg.com',
-    atype: 1
-  },
-
-  // pairId
-  'pairId': {
-    source: 'google.com',
-    atype: 571187
-  },
-
-  // justId
-  'justId': {
-    source: 'justtag.com',
-    atype: 1
-  },
-
-  // pubCommonId
-  'pubcid': {
-    source: 'pubcid.org',
-    atype: 1
-  },
-
-  // unifiedId
-  'tdid': {
-    source: 'adserver.org',
-    atype: 1,
-    getUidExt: function() {
-      return {
-        rtiPartner: 'TDID'
-      };
-    }
-  },
-
-  // id5Id
-  'id5id': {
-    getValue: function(data) {
-      return data.uid
-    },
-    source: 'id5-sync.com',
-    atype: 1,
-    getUidExt: function(data) {
-      if (data.ext) {
-        return data.ext;
-      }
-    }
-  },
-
-  // ftrack
-  'ftrackId': {
-    source: 'flashtalking.com',
-    atype: 1,
-    getValue: function(data) {
-      let value = '';
-      if (data && data.ext && data.ext.DeviceID) {
-        value = data.ext.DeviceID;
-      }
-      return value;
-    },
-    getUidExt: function(data) {
-      return data && data.ext;
-    }
-  },
-
-  // parrableId
-  'parrableId': {
-    source: 'parrable.com',
-    atype: 1,
-    getValue: function(parrableId) {
-      if (parrableId.eid) {
-        return parrableId.eid;
-      }
-      if (parrableId.ccpaOptout) {
-        // If the EID was suppressed due to a non consenting ccpa optout then
-        // we still wish to provide this as a reason to the adapters
-        return '';
-      }
-      return null;
-    },
-    getUidExt: function(parrableId) {
-      const extendedData = pick(parrableId, [
-        'ibaOptout',
-        'ccpaOptout'
-      ]);
-      if (Object.keys(extendedData).length) {
-        return extendedData;
-      }
-    }
-  },
-
-  // identityLink
-  'idl_env': {
-    source: 'liveramp.com',
-    atype: 3
-  },
-
-  // liveIntentId
-  'lipb': {
-    getValue: function(data) {
-      return data.lipbid;
-    },
-    source: 'liveintent.com',
-    atype: 3,
-    getEidExt: function(data) {
-      if (Array.isArray(data.segments) && data.segments.length) {
-        return {
-          segments: data.segments
-        };
-      }
-    }
-  },
-
-  // bidswitchId
-  'bidswitch': {
-    source: 'bidswitch.net',
-    atype: 3,
-    getValue: function(data) {
-      return data.id;
-    },
-    getUidExt: function(data) {
-      if (data.ext) {
-        return data.ext;
-      }
-    }
-  },
-
-  // medianetId
-  'medianet': {
-    source: 'media.net',
-    atype: 3,
-    getValue: function(data) {
-      return data.id;
-    },
-    getUidExt: function(data) {
-      if (data.ext) {
-        return data.ext;
-      }
-    }
-  },
-
-  // magnite
-  'magnite': {
-    source: 'rubiconproject.com',
-    atype: 3,
-    getValue: function(data) {
-      return data.id;
-    },
-    getUidExt: function(data) {
-      if (data.ext) {
-        return data.ext;
-      }
-    }
-  },
-
-  // britepoolId
-  'britepoolid': {
-    source: 'britepool.com',
-    atype: 3
-  },
-
-  // dmdId
-  'dmdId': {
-    source: 'hcn.health',
-    atype: 3
-  },
-
-  // lotamePanoramaId
-  lotamePanoramaId: {
-    source: 'crwdcntrl.net',
-    atype: 1,
-  },
-
-  // criteo
-  'criteoId': {
-    source: 'criteo.com',
-    atype: 1
-  },
-
-  // merkleId
-  'merkleId': {
-    atype: 3,
-    getSource: function(data) {
-      if (data?.ext?.ssp) {
-        return `${data.ext.ssp}.merkleinc.com`
-      }
-      return 'merkleinc.com'
-    },
-    getValue: function(data) {
-      return data.id;
-    },
-    getUidExt: function(data) {
-      if (data.keyID) {
-        return {
-          keyID: data.keyID
-        }
-      }
-      if (data.ext) {
-        return data.ext;
-      }
-    }
-  },
-
-  // NetId
-  'netId': {
-    source: 'netid.de',
-    atype: 1
-  },
-
-  // zeotapIdPlus
-  'IDP': {
-    source: 'zeotap.com',
-    atype: 1
-  },
-
-  // hadronId
-  'hadronId': {
-    source: 'audigent.com',
-    atype: 1
-  },
-
-  // quantcastId
-  'quantcastId': {
-    source: 'quantcast.com',
-    atype: 1
-  },
-
-  // IDx
-  'idx': {
-    source: 'idx.lat',
-    atype: 1
-  },
-
-  // Verizon Media ConnectID
-  'connectid': {
-    source: 'verizonmedia.com',
-    atype: 3
-  },
-
-  // Neustar Fabrick
-  'fabrickId': {
-    source: 'neustar.biz',
-    atype: 1
-  },
-
-  // MediaWallah OpenLink
-  'mwOpenLinkId': {
-    source: 'mediawallahscript.com',
-    atype: 1
-  },
-
-  'tapadId': {
-    source: 'tapad.com',
-    atype: 1
-  },
-
-  // Novatiq Snowflake
-  'novatiq': {
-    getValue: function(data) {
-      if (data.snowflake.id === undefined) {
-        return data.snowflake;
-      }
-
-      return data.snowflake.id;
-    },
-    source: 'novatiq.com',
-  },
-
-  'uid2': {
-    source: 'uidapi.com',
-    atype: 3,
-    getValue: function(data) {
-      return data.id;
-    },
-    getUidExt: function(data) {
-      if (data.ext) {
-        return data.ext;
-      }
-    }
-  },
-
-  'euid': {
-    source: 'euid.eu',
-    atype: 3,
-    getValue: function(data) {
-      return data.id;
-    }
-  },
-
-  'deepintentId': {
-    source: 'deepintent.com',
-    atype: 3
-  },
-
-  // Admixer Id
-  'admixerId': {
-    source: 'admixer.net',
-    atype: 3
-  },
-
-  // Adtelligent Id
-  'adtelligentId': {
-    source: 'adtelligent.com',
-    atype: 3
-  },
-
-  amxId: {
-    source: 'amxdt.net',
-    atype: 1,
-  },
-
-  'publinkId': {
-    source: 'epsilon.com',
-    atype: 3
-  },
-
-  'kpuid': {
-    source: 'kpuid.com',
-    atype: 3
-  },
-
-  'imppid': {
-    source: 'ppid.intimatemerger.com',
-    atype: 1
-  },
-
-  'imuid': {
-    source: 'intimatemerger.com',
-    atype: 1
-  },
-
-  // Yahoo ConnectID
-  'connectId': {
-    source: 'yahoo.com',
-    atype: 3
-  },
-
-  // Adquery ID
-  'qid': {
-    source: 'adquery.io',
-    atype: 1
-  },
-
-  // DAC ID
-  'dacId': {
-    source: 'impact-ad.jp',
-    atype: 1
-  },
-
-  // 33across ID
-  '33acrossId': {
-    source: '33across.com',
-    atype: 1,
-    getValue: function(data) {
-      return data.envelope;
-    }
-  },
-
-  // tncId
-  'tncid': {
-    source: 'thenewco.it',
-    atype: 3
-  },
-
-  // Gravito MP ID
-  'gravitompId': {
-    source: 'gravito.net',
-    atype: 1
-  },
-
-  // czechAdId
-  'czechAdId': {
-    source: 'czechadid.cz',
-    atype: 1
-  },
-
-  // OneKey Data
-  'oneKeyData': {
-    getValue: function(data) {
-      if (data && Array.isArray(data.identifiers) && data.identifiers[0]) {
-        return data.identifiers[0].value;
-      }
-    },
-    source: 'paf',
-    atype: 1,
-    getEidExt: function(data) {
-      if (data && data.preferences) {
-        return {preferences: data.preferences};
-      }
-    },
-    getUidExt: function(data) {
-      if (data && Array.isArray(data.identifiers) && data.identifiers[0]) {
-        const id = data.identifiers[0];
-        return {
-          version: id.version,
-          type: id.type,
-          source: id.source
-        };
-      }
-    }
-  }
-};
+export const USER_IDS_CONFIG = {};
 
 // this function will create an eid object for the given UserId sub-module
 function createEidObject(userIdData, subModuleKey) {
